Moved makeDeploymentList to transformer
