+++ veteran-verification-api/src/test/java/gov/va/api/lighthouse/veteranverification/api/PayGradeTest.java
@@ -6,6 +6,21 @@ import org.junit.jupiter.api.Test;
 
 public class PayGradeTest {
   @Test
+  public void happyPath() {
+    String expected = "E05";
+    String actual = PayGrade.builder().payPlanCode("ME").payGradeCode("05").build().toString();
+    Assertions.assertEquals(expected, actual);
+  }
+
+  @Test
+  public void happyPathExtraSpace() {
+    String expected = "E05";
+    String actual =
+        PayGrade.builder().payPlanCode("   ME ").payGradeCode("  05   ").build().toString();
+    Assertions.assertEquals(expected, actual);
+  }
+
+  @Test
   public void payGradeCodeIsEmpty() {
     String expected = "unknown";
     String actual = PayGrade.builder().payPlanCode("ME").payGradeCode("").build().toString();
@@ -30,21 +45,6 @@ public class PayGradeTest {
   public void payPlanCodeIsNull() {
     String expected = "unknown";
     String actual = PayGrade.builder().payPlanCode(null).payGradeCode("05").build().toString();
-    Assertions.assertEquals(expected, actual);
-  }
-
-  @Test
-  public void happyPath() {
-    String expected = "E05";
-    String actual = PayGrade.builder().payPlanCode("ME").payGradeCode("05").build().toString();
-    Assertions.assertEquals(expected, actual);
-  }
-
-  @Test
-  public void happyPathExtraSpace() {
-    String expected = "E05";
-    String actual =
-        PayGrade.builder().payPlanCode("   ME ").payGradeCode("  05   ").build().toString();
     Assertions.assertEquals(expected, actual);
   }
 }
